make ComplexColumn an interface and ExtensionPoint
