+++ processing/src/main/java/org/apache/druid/segment/column/ComplexColumn.java
@@ -19,45 +19,29 @@
 
 package org.apache.druid.segment.column;
 
+import org.apache.druid.guice.annotations.ExtensionPoint;
 import org.apache.druid.query.monomorphicprocessing.RuntimeShapeInspector;
 import org.apache.druid.segment.ColumnValueSelector;
 import org.apache.druid.segment.ObjectColumnSelector;
-import org.apache.druid.segment.data.GenericIndexed;
 import org.apache.druid.segment.data.ReadableOffset;
 
 import javax.annotation.Nullable;
 
 /**
-*/
-public class ComplexColumn implements BaseColumn
+ */
+@ExtensionPoint
+public interface ComplexColumn extends BaseColumn
 {
-  private final GenericIndexed<?> index;
-  private final String typeName;
-
-  public ComplexColumn(String typeName, GenericIndexed<?> index)
-  {
-    this.index = index;
-    this.typeName = typeName;
-  }
-
-  public String getTypeName()
-  {
-    return typeName;
-  }
-
-  @Nullable
-  public Object getRowValue(int rowNum)
-  {
-    return index.get(rowNum);
-  }
-
-  public int getLength()
-  {
-    return index.size();
-  }
+  Class<?> getClazz();
+  String getTypeName();
+  Object getRowValue(int rowNum);
+  int getLength();
 
   @Override
-  public ColumnValueSelector<?> makeColumnValueSelector(ReadableOffset offset)
+  void close();
+
+  @Override
+  default ColumnValueSelector makeColumnValueSelector(ReadableOffset offset)
   {
     return new ObjectColumnSelector()
     {
@@ -71,7 +55,7 @@ public class ComplexColumn implements BaseColumn
       @Override
       public Class classOfObject()
       {
-        return index.getClazz();
+        return getClazz();
       }
 
       @Override
@@ -80,11 +64,5 @@ public class ComplexColumn implements BaseColumn
         inspector.visit("column", ComplexColumn.this);
       }
     };
-  }
-
-  @Override
-  public void close()
-  {
-    // nothing to close
   }
 }
